create isolate component slide for template review app
